Move while refactor
