+++ src/index.js
@@ -1,6 +1,6 @@
 import './styles.css';
 
-import * as Game from './Game'
+import * as Game from './refactored/Game'
 
 if (module.hot) {
   module.hot.accept(function () {
