+++ src/refactored/Entites.js
