+++ src/refactored/Renderer.js
